Move handleError and extractData to services
